Exercicio modulo 9, alteração

+++ BACK-END/modulo-9/tarefa_wrapper/Main.java
@@ -2,7 +2,7 @@ public class Main {
     public static void main(String[] args) {
 
         long num1 = 22222222222222l;
-        int num2 = (int) num1;
+        Long num2 = (Long) num1;
 
         System.out.println(num2);
